requirements.txt
# Stock Predictor Dependencies

# Neural network framework
torch>=2.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Massive (formerly Polygon.io) API client
polygon-api-client>=1.13.0

# Anthropic API for sentiment analysis
anthropic>=0.25.0
